PP-8443 Get Webhook resource
Adds ability to get webhook by external id.

Controversial things what I have done:

- renamed webhook entity to `Webhook` and used
this both as jpa entity and jackson response.
I think domain is currently sufficiently simple
that this is ok, and I reallly don't like writing
lots of things to change stuff from one pojo to
another. Happy to hear other opinions

- used named typed query and defined this on
Webhook entity. To my eye this makes everything
a bit cleaner and I like that queries are defined
with entity, and then DAO is responsible for
executing and handling result. Open to discuss this.

- removed mocking of Webhook thing from resource test:
this was breaking Jackson serialisation, and seems
unnecessary

- written one big WebhookResourceIT that posts to API
then retrieves from API. Seems like a nice test to have for now
and avoids need to squirt things in to DB to set up. May not
be sustainable.