PP-8557: ignore pending delivery queue items for message status
The most up to date message status should filter for terminal statuses
that have already happened. Pending statuses are useful for a detailed
view but don't reflect what has already happened on any given message.

Reflect this on the latest (but past tense) status on the message delivery
queueitem status.