PP-8557: restrict list webhook message status filter to valid statuses
If a `status` query param is provided, make sure it is one that can be
serialised/ deserialised into the database.

If none is provided allow the query through.

Update queries to do a direct equality with a single status rather than
anticipate a list of potentially multiple statuses to filter.

Clean up optional behaviour favouring idiomatic Java.