BAU: add basic indexes on webhook tables
Use the queries defined in the entities to cover attributes used in gets
and searches with indexes.
- `WebhookEntity`
- `WebhookMessageEntity`
- `WebhookDeliveryQueueEntity`

Also include a unique index constraint on the subscription table to both
speed to the many to one join and to avoid code mistakenly subscribing a
webhook to any given event hook more than once.