BAU: fix invalid liquibase file
If liquibase picks up an sql file with no comments it will create a
default `changeset` (in this example
`changeset:0008_index_searched_columns.sql`).

If there is a liquibase comment but no explicit `changeset` comment it
looks like the file is not run by liquibase as it doesn't apply the
default behaviour.

Remove comments to use the default liquibase changeset and rollback
behaviour for this change.