BAU: replace internal hibernate package with dropwizard unit of work
The current solution for ensuring a valid db transaction context is to
use hibernate `internal` methods to open and bind a session manually.

This is required because dropwizard managed `UnitOrWork` contexts are
only available within Jersey resources by default.

Use the unit of work aware proxy factory to build a polling service that
is wrapped in the dropwizard hibernate managed context and can make use
of the standard `@UnitOfWork` annotations.

This refactor follows a pattern set in
https://github.com/alphagov/pay-webhooks/pull/251