PP-9016 Check domain is in allow list before sending message
Check that the domain of a webhook callback URL is in the allow
list before sending a webhook message.

The allow list is a simple list of domains like gov.uk and
example.com. When sending a message, we check if the domain is in
the allow list. Subdomains are allowed, so if the allow list
contains gov.uk and example.com, we will allow callback URLs with
domains like www.service.gov.uk and www.example.com. Since domain
names are case-insensitive, so is the check (for plain ASCII
domain names).

The implementation relies on Guava’s InternetDomainName class.
Each domain name on the allow list is converted into an
InternetDomainName. We then convert the callback URL to a URL
object (we chose this over an URI object because URL is more
specific for web addresses and supports international domain
names) and extract the hostname. If the URL has a host name
and it’s a domain name (rather than an IP address), we convert
it into an InternetDomainName (if it neither of these things
are true, we disallow the sending of the webhook message). We then
check if the InternetDomainName in is in the allow list. If it is,
we will allow the webhook message to be sent. If it is not, we lop
the leftmost part of the domain name off (for example,
www.service.gov.uk becomes service.gov.uk) to create a new
InternetDomainName and then look for that in the allow list. We
continue doing this until there are just two parts of the domain
left and do a final check on that (this means a single-part domain
name like com will never be found in the allow list). If we have
not found the domain name in the allow list by this point, we do
not allow the webhook message to be sent.

We made a choice to rely heavily on the InternetDomainName class
to avoid having to deal with domain name weirdness ourselves. An
earlier draft of this code kept the allow list as strings and
avoided creating a new InternetDomainName each time a part is
lopped off the domain name (with allow its associated parsing and
validation) but testing showed this was only roughly three times
faster, which did not seem a big enough win when considered against
the simplicity of just using InternetDomainName.

International domain names are supported by the check does not
consider IDNs and their Punycode equivalents to match. If we ever
have to add an IDN to the whitelist, we could add the Punycode
equivalent as well to ensure both work.

with @alexbishop1