BAU: disable hibernate statistics logs
Hibernate seems to be generating and logging statistics for every single
database connection, by default.

How this behvaiour was introduced or is on by default isn't clear at
this stage, however with production volumes the logs generated are
problematic.

Disable these logs by setting the hibernate property and investigate
why this behaviour is enabled in the first place.