PP-10466: fix webhook delivery queue next to send query
The webhook delivery queue DAO exposes a method to search for webhook
messages which are in a pending state and are now scheduled to be sent.

There are a number of individual indexes on the delivery queue table:
- `send_at`
- `created_at`
- `delivery_status`

Both `send_at` and `delivery_status` are used in the query, which should
correctly use index scans to immediately return the value.

An order (through the named query) and a limit (through a
hibernate setter) are also applied to the query, by order by `send_at`
and limting to 1 we are asking postgres to use both the `send_at` and
`delivery_status` index at the same time which results in no index being
used -- if we wanted to include this ordering we would need to have a
compoud index between these two columns.

However it doesn't look like the ordering for which message to send
should be too important in this case, all messages that meet these
criteria will need to be immediately attempted and for a healthy emitter
system this should never see a significant backlog. The order will now
come from the default read order which will likely favour newly created
messages which feels appropriate.

By removing the order, the indexes are individually used and the query
should immediately execute as expected.