PP-10631: Specify URL when logging 'Sending webhook message finished' (#626)
* PP-10631: Specify URL when logging 'Sending webhook message finished'

As part of protective monitoring, there is an alert to compare webhooks and
webhooks-egress target domains. Currently the alert's search on the webhooks
microservice logs is searching for "Sending webhook message". However this isn't correct as
the presence of this log doesn't indicate webhooks microservice actually did
call out to the target url. This is because at the point of this log, the
target webhook url may be inactive, not on the allow list or failed because of
some HTTP failure (e.g. socket timeout, interrupted exception). The splunk
alert should search where the log has "Sending webhook message finished" and
"to_state" is "SUCCESSFUL".

This commit adds the "domain" field to where "Sending webhook message finished"
is logged so splunk can extract the log properly.