PP-10543: abstract SQS events fixture
When a SQS message is processed by webhooks the message was initially
fanned out from SNS. The shape and properties of this message are
largely static (as we're currenlty only dealing with card events) with a
variable `"Message"` parameter which includes the original event body.

Clean up existing tests by abstracting out the SNS schema details into a
fixture that allows specifying a simple map for the event body (the
fixture builder will deal with serialisation, JSON in JSON etc.).

This should let us use this fixture in other unit/ integration tests
without duplicated schema.