PP-10543: filter webhook messages by last delivery status
Instead of relying on the delivery queue status for the last attempt,
directly use the `lastDeliveryStatus` property now that it exists and is
kept up to date.

This will allow JPA to lazy load the last delivery queue which will
significatly decrease the complexity of the query for fetching the
messages based on status.

Tests for filtering/ paginating webhook messages should continue to
pass.