PP-10543: update get webhook message query to use webhook entity (ID)
Update the get one webhook message query to directly reference the
webhook entity (using the internal identifier) rather than relying on a
sub query or a join.

Using postgres `explain` and `explain analyse` we've observed that using
the external ID from the right hand side of the join or adding a sub
query to lookup the webhook ID from an external ID both cause postgres
to not use appropriate indexes and do a full (reverse) table scan on the
webhook messages table.

As we don't need the full webhoook details in this instance, set the
fetch mode to lazy and directly reference the webhook entity to ensure
appropriate indexes are used for the query.

In order to do this cleanly, this commit also resolves an issue where
the previous get message DAO method was not returning an optional, which
meant an empty result would cause a null pointer exception when being
mapped to a JSON response. By making this optional at the DAO layer,
each subsequent layer (service then resource) handle the case of there
being no data appropriately. Add additional coverage to make sure we're
correctly mapping the get message resource to a 404.