PP-10537: add override flag to get one webhook endpoint
Internal tools like Toolbox internally consume this resource and might
be trying to access a webhook outside of an account/ service context.
The standard pattern for this on other endpoints is to require the
identifier to be provided (to avoid internal code loading data by
mistake) but to allow internal tools to provice the
`override_account_or_service_id_restriction` flag.

Check existing resource tests pass and add a scenario for providing the
flag without providing a service identifier.