PP-10835: Dao tests for deleting, and getting rows by date
As we want to remove webhooks messages based on a maximum age of message,
introduce methods to retrieve the WebhookMessageEntities and
WebhookDeliveryQueueEntities based on date, and methods to delete them. These
are just simple Dao tests to ensure the database access code is correct. More
complex integration tests will be introduced in a later commit.