BAU: reduce initial thread delay for sending webhook messages
There is currently a 30 second default delay before the webhook message
emitter background process starts sending messages.

There is a hypothesis that this can lead to a "healthy" app deployed to
an evironment but then disrupt tests for either that app or other apps
under test by being configured to not send messages for another 30 seconds.

As the SQS and RDS instances for webhooks are not turned on or off
regularly, the app should be able to directly connect, there should be
no issues with competing threads with these resources.

The healthcheck resource is also responsible for ensuring that before http
traffic is routed to the app, it can:
- respond to HTTP requests
- connect to the database
- connect to SQS

Note this doesn't stop background processes from starting up.