PP-11019: add gateway account id to webhooks table
Adds a new column and index to the webhooks table. The column is
specified as varchar to line up with how ledger stores and represents
this data.

Create the index concurrently (and only if it doesn't exist) to avoid
locking the table for a long period of time if there is signficant data
in the environment.