PP-9850: fix serialising child events
Currently webhooks only supports sending events that relate to payments.
Early thinking on how to populate the resource on a webhook message
decided to format the resource according to the event type, however this
won't always work (in this case we want to send the payment that has
been refunded, rather than the refund).

This removes that logic but leaves the tools in place in case we need to
allow services to subscribe to different types of resources in the
future (refunds, disputes, agreemenets, etc.)