ZD-5392584: include payment expired event
A user has observed that payments that expire without any interaction
don't receive the webhook message for "Payment expired".

It looks like the event stream differentiates between payments that are
expired and require cancelling (have already authorised the payment with
the gateway) and payments that are able to be immediately expired
without any action.

Include the default expired event as part of the "Payment expired"
subscription to allow services to be updated when this default case
occurs.