PP-11406: Log domain even if request timed out (#958)
If the request timed out, the request was still attempted. If the request was
attempted, the domain is logged in the egress logs. Currently, if the reqest is
timed out, the domain isn't logged in the application logs, which leads to a
mismatch between the domains logged in the webhook application logs and egress
logs. This causes the "Compare webhooks and webhooks-egress target domains"
alert to fire unnecessarily. This commit logs the domain even if the request
times out, and uses `URI.create` which does not throw a checked exception which
makes the code a bit nicer.