bau: Revert to using java.net.URL
It turns out that we used java.net.URL because of InternetDomainName support -
see https://github.com/alphagov/pay-webhooks/pull/362.

Also strip leading and trailing whitespaces in the callback url. There is
currently a url with a trailing whitespace. This is causing the current
`URI.create` to throw a RuntimeException which is causing our sentry budget to
be eaten up.